Removing the RST pin and releasing the MicroSD.
I discovered a way to make the Si4713 (FM Transmitter) work without the RST pin plugged into the Cardputer and StcikC. It is necessary to take the cable connected to the RST pin of the FM Transmitter and quickly touch it to the GND pin, changing the state of the pin and activating the board. In the future it is possible to add some hardware to do this work.


Removing the RST pin from the code also allows the use of the MicroSD Card from the SD Card Sniffer or Cardputer, improving usability.